Set HSF_PLATFORM to reasonable default
As reported in Issue #6, template CPack file relies on HSF program
hsf_get_platform.py being present in the path. If it is not, the
platform variable HSF_PLATFORM is unset.

If HSF_PLATFORM is empty after an attempt to run hsf_get_platform.py,
set it to a value derived from CMake's knowledge of the system
processor arch, name and CXX compiler vendor ID. Do not use system
or compiler versions as these do not match the HSF conventions (e.g.
system version on Linux/Darwin is the kernel version).

Note that this assumes a CXX compiler has been configured.