Use explicit source listing
GLOBs in CMake are only evaluated by cmake and not the buildtool
generated for. With GLOB, adding/removing files requires an explicit,
by hand rerun of cmake in order to add/remove the files from the
buildscripts. This can lead to inconsistent builds and/or cryptic
error messages especially when developing over VCS. GLOBs also make
it more complicated to handle configurable builds where sources may be
used or not depending on configure-time options.

Explicit listing overcomes these issues as the CMake scripts are a
dependency of the project. Adding/removing a file and its listing in
a CMake script result it an automatic re-run of CMake before the build
itself is rerun. Forgetting to add/remove a file listing typically
leads to faster errors, with clearer error messages.

Remove use of GLOB to list source files, following CMake recommended
practice. Projects may use GLOB in the future, but HSF should not
recommend it in the first instance.